api/src/requirements.txt
psycopg2-binary >=2.9.9, <3.0.0
faker >=19.12.0, <20.0.0
fastapi >=0.104.1, <0.105.0
kafka-python >=2.0.2, <2.1.0
uvicorn >=0.23.2, <0.24.0
python-dotenv >=1.0.0, <2.0.0
pytz >=2023.3.post1, <2024.0.0
pydantic[email] >=2.4.2, <2.5.0
pandas >=2.1.2, <2.2.0
boto3 >=1.28.75, <1.29.0
pyarrow >=14.0.0, <15.0.0